# project/docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Auditable Agent Runtime API
  version: 1.0.0
  description: Minimum MVP contract for a multi-tenant auditable Agent Runtime.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /auth/refresh:
    post:
      security: []
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200': { description: Tokens refreshed, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /workflows:
    get:
      operationId: listWorkflows
      responses:
        '200': { description: Tenant workflow definitions }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /workflows/{workflow_id}:
    get:
      operationId: getWorkflow
      parameters: [{ name: workflow_id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Workflow definition and recent instances }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateWorkflow
      parameters: [{ name: workflow_id, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [nodes, edges], properties: { nodes: { type: array, maxItems: 100, items: { type: object } }, edges: { type: array, maxItems: 200, items: { type: object } } } } } } }
      responses:
        '200': { description: Workflow draft saved }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /workflows/{workflow_id}/duplicate:
    post:
      summary: Duplicate workflow
      description: Creates a copy of an existing workflow.
      operationId: duplicateWorkflow
      parameters: [{ name: workflow_id, in: path, required: true, schema: { type: string } }]
      responses:
        '201': { description: Workflow duplicated }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: opaque }
...
  responses:
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Forbidden: { description: Action is not allowed, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    ValidationError: { description: Invalid request, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
  schemas:
...
    RefreshRequest:
      type: object
      required: [refreshToken]
      additionalProperties: false
      properties:
        refreshToken: { type: string, minLength: 32, maxLength: 256, writeOnly: true }
    AuthResponse:
      type: object
      required: [data, requestId]
      properties:
        data: { type: object, required: [accessToken, expiresAt], properties: { accessToken: { type: string, writeOnly: true }, expiresAt: { type: string, format: date-time } } }
        requestId: { type: string }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string, enum: [UNAUTHORIZED, FORBIDDEN, NOT_FOUND, VALIDATION_ERROR, CONFLICT, VERSION_CONFLICT, RATE_LIMIT_EXCEEDED, QUOTA_EXCEEDED, BUDGET_EXCEEDED, INVALID_STATE_TRANSITION, APPROVAL_ALREADY_HANDLED, RESUME_WINDOW_EXPIRED, MODEL_PROVIDER_ERROR, TOOL_EXECUTION_FAILED, SANDBOX_TIMEOUT, INTERNAL_ERROR] }
            message: { type: string, maxLength: 500 }
            requestId: { type: string }
            details: { type: object, additionalProperties: true }
...